MongoDB robust

+++ app.py
@@ -17,40 +17,45 @@ load_dotenv()
 app = Flask(__name__)
 
 # Testing Parameters
-TEST_CONNECTION = False     # ? Set to True to test connection to MongoDB
-REAL_DATA = False            # ? Set to False to use sample data
+REAL_DATA = False            # ? False: use sample data / True: use real data
+TEST_CONNECTION = False     # ? False: don't test connection / True: test connection
 
 
-# MongoDB connection
-mongodb_uri = os.getenv("MONGODB_URI")
-client = MongoClient(mongodb_uri, server_api=ServerApi('1'))
+# Define db and collection as placeholders
+db = None
+collection = None
 
-# Select the database and collection
-db = client['DTL_db']
-collection = db['posts']
+if REAL_DATA:
+    # MongoDB connection
+    mongodb_uri = os.getenv("MONGODB_URI")
+    client = MongoClient(mongodb_uri, server_api=ServerApi('1'))
 
-# Send a ping to confirm a successful connection
-if TEST_CONNECTION:
-    try:
-        client.admin.command('ping')
-        print("✅ Pinged your deployment. You successfully connected to MongoDB!")
-    except Exception as e:
-        print(e)
+    # Select the database and collection
+    db = client['DTL_db']
+    collection = db['posts']
 
-    # Add a sample document in the collection (don't add if already exists)
-    name = 'John Doe'
-    if collection.count_documents({'name': name}) == 0:
-        collection.insert_one({'name': name})
-        print("✅ Added a sample document in the collection.")
+    # Send a ping to confirm a successful connection
+    if TEST_CONNECTION:
+        try:
+            client.admin.command('ping')
+            print("✅ Pinged your deployment. You successfully connected to MongoDB!")
+        except Exception as e:
+            print(e)
 
-    # Retrieve the sample document added
-    result = collection.find_one({'name': name})
-    print(result)
-    print("✅ Retrieved the sample document.")
+        # Add a sample document in the collection (don't add if already exists)
+        name = 'John Doe'
+        if collection.count_documents({'name': name}) == 0:
+            collection.insert_one({'name': name})
+            print("✅ Added a sample document in the collection.")
 
-    # Delete the sample document added
-    collection.delete_one({'name': name})
-    print("✅ Deleted the sample document.")
+        # Retrieve the sample document added
+        result = collection.find_one({'name': name})
+        print(result)
+        print("✅ Retrieved the sample document.")
+
+        # Delete the sample document added
+        collection.delete_one({'name': name})
+        print("✅ Deleted the sample document.")
 
 
 @app.route('/')
